GeneratedRouter to its own component file
