+++ frontend-react/src/components/Content/MarkdownDirectory.tsx
@@ -1,9 +1,5 @@
 import * as module from "module";
 
-import { Route } from "react-router-dom";
-
-import DirectoryAsPage from "./DirectoryAsPage";
-
 type ContentElement = () => JSX.Element;
 
 /** A base type that holds directory information
@@ -62,45 +58,3 @@ export class MarkdownDirectory extends ContentDirectory {
         return this;
     }
 }
-/** Takes a `ContentDirectory` and returns a react-router `Route` */
-export const GeneratedRoute = ({ dir }: { dir: ContentDirectory }) => {
-    if (dir instanceof MarkdownDirectory) {
-        return (
-            <Route
-                key={`${dir.slug}-route`}
-                path={`${dir.slug}`}
-                element={
-                    <DirectoryAsPage
-                        key={`${dir.slug}-dir-as-page`}
-                        directory={dir}
-                    />
-                }
-            />
-        );
-    } else {
-        const castDir = dir as ElementDirectory;
-        return (
-            <Route
-                key={`${castDir.slug}-route`}
-                path={`${castDir.slug}`}
-                element={castDir.element}
-            />
-        );
-    }
-};
-/** Takes a `ContentDirectory[]` and generates a React Fragment containing
- * each directory's `GeneratedRoute` */
-export const GeneratedRouter = ({
-    directories,
-}: {
-    directories: ContentDirectory[];
-}) => {
-    return (
-        <>
-            {directories.map((dir, idx) => (
-                <GeneratedRoute dir={dir} key={idx} />
-            ))}
-            {/* Handles any undefined route */}
-        </>
-    );
-};
